Move gencopy to generational::copying
