+++ src/plan/generational/mod.rs
@@ -1,1 +1,4 @@
 pub mod global;
+pub mod gc_work;
+
+pub mod copying;
