+++ src/plan/global.rs
@@ -123,7 +123,7 @@ pub fn create_mutator<VM: VMBinding>(
         PlanSelector::SemiSpace => {
             crate::plan::semispace::mutator::create_ss_mutator(tls, &*mmtk.plan)
         }
-        PlanSelector::GenCopy => crate::plan::gencopy::mutator::create_gencopy_mutator(tls, mmtk),
+        PlanSelector::GenCopy => crate::plan::generational::copying::mutator::create_gencopy_mutator(tls, mmtk),
         PlanSelector::MarkSweep => {
             crate::plan::marksweep::mutator::create_ms_mutator(tls, &*mmtk.plan)
         }
@@ -147,7 +147,7 @@ pub fn create_plan<VM: VMBinding>(
             vm_map, mmapper, options,
         )),
         PlanSelector::GenCopy => {
-            Box::new(crate::plan::gencopy::GenCopy::new(vm_map, mmapper, options))
+            Box::new(crate::plan::generational::copying::GenCopy::new(vm_map, mmapper, options))
         }
         PlanSelector::MarkSweep => Box::new(crate::plan::marksweep::MarkSweep::new(
             vm_map, mmapper, options,
